+++ src/PhpSpreadsheet/Worksheet/Worksheet.php
@@ -4,7 +4,11 @@ namespace PhpOffice\PhpSpreadsheet\Worksheet;
 
 use ArrayObject;
 use PhpOffice\PhpSpreadsheet\Calculation\Calculation;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Cell\AddressRange;
 use PhpOffice\PhpSpreadsheet\Cell\Cell;
+use PhpOffice\PhpSpreadsheet\Cell\CellAddress;
+use PhpOffice\PhpSpreadsheet\Cell\CellRange;
 use PhpOffice\PhpSpreadsheet\Cell\Coordinate;
 use PhpOffice\PhpSpreadsheet\Cell\DataType;
 use PhpOffice\PhpSpreadsheet\Cell\DataValidation;
@@ -1104,22 +1108,118 @@ class Worksheet implements IComparable
     }
 
     /**
+     * Validate a cell address.
+     *
+     * @param null|array<int>|CellAddress|string $cellAddress Coordinate of the cell as a string, eg: 'C5';
+     *               or as an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
+     */
+    protected function validateCellAddress($cellAddress): string
+    {
+        if (is_string($cellAddress)) {
+            [$worksheet, $address] = self::extractSheetTitle($cellAddress, true);
+//            if (!empty($worksheet) && $worksheet !== $this->getTitle()) {
+//                throw new Exception('Reference is not for this worksheet');
+//            }
+
+            return empty($worksheet) ? strtoupper($address) : $worksheet . '!' . strtoupper($address);
+        }
+
+        if (is_array($cellAddress)) {
+            $cellAddress = CellAddress::fromColumnRowArray($cellAddress);
+        }
+
+        return (string) $cellAddress;
+    }
+
+    private function tryDefinedName(string $coordinate): string
+    {
+        // Uppercase coordinate
+        $coordinate = strtoupper($coordinate);
+        // Eliminate leading equal sign
+        $coordinate = self::pregReplace('/^=/', '', $coordinate);
+        $defined = $this->parent->getDefinedName($coordinate, $this);
+        if ($defined !== null) {
+            if ($defined->getWorksheet() === $this && !$defined->isFormula()) {
+                $coordinate = self::pregReplace('/^=/', '', $defined->getValue());
+            }
+        }
+
+        return $coordinate;
+    }
+
+    /**
+     * Validate a cell address or cell range.
+     *
+     * @param AddressRange|array<int>|CellAddress|int|string $cellRange Coordinate of the cells as a string, eg: 'C5:F12';
+     *               or as an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 12]),
+     *               or as a CellAddress or AddressRange object.
+     */
+    protected function validateCellOrCellRange($cellRange): string
+    {
+        if (is_string($cellRange) || is_numeric($cellRange)) {
+            $cellRange = (string) $cellRange;
+            // Convert a single column reference like 'A' to 'A:A'
+            $cellRange = self::pregReplace('/^([A-Z]+)$/', '${1}:${1}', $cellRange);
+            // Convert a single row reference like '1' to '1:1'
+            $cellRange = self::pregReplace('/^(\d+)$/', '${1}:${1}', $cellRange);
+        } elseif (is_object($cellRange) && $cellRange instanceof CellAddress) {
+            $cellRange = new CellRange($cellRange, $cellRange);
+        }
+
+        return $this->validateCellRange($cellRange);
+    }
+
+    /**
+     * Validate a cell range.
+     *
+     * @param AddressRange|array<int>|string $cellRange Coordinate of the cells as a string, eg: 'C5:F12';
+     *               or as an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 12]),
+     *               or as an AddressRange object.
+     */
+    protected function validateCellRange($cellRange): string
+    {
+        if (is_string($cellRange)) {
+            [$worksheet, $addressRange] = self::extractSheetTitle($cellRange, true);
+
+            // Convert Column ranges like 'A:C' to 'A1:C1048576'
+            $addressRange = self::pregReplace('/^([A-Z]+):([A-Z]+)$/', '${1}1:${2}1048576', $addressRange);
+            // Convert Row ranges like '1:3' to 'A1:XFD3'
+            $addressRange = self::pregReplace('/^(\\d+):(\\d+)$/', 'A${1}:XFD${2}', $addressRange);
+
+            return empty($worksheet) ? strtoupper($addressRange) : $worksheet . '!' . strtoupper($addressRange);
+        }
+
+        if (is_array($cellRange)) {
+            [$from, $to] = array_chunk($cellRange, 2);
+            $cellRange = new CellRange(CellAddress::fromColumnRowArray($from), CellAddress::fromColumnRowArray($to));
+        }
+
+        return (string) $cellRange;
+    }
+
+    /**
      * Set a cell value.
      *
-     * @param string $coordinate Coordinate of the cell, eg: 'A1'
-     * @param mixed $value Value of the cell
+     * @param array<int>|CellAddress|string $coordinate Coordinate of the cell as a string, eg: 'C5';
+     *               or as an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
+     * @param mixed $value Value for the cell
      *
      * @return $this
      */
     public function setCellValue($coordinate, $value)
     {
-        $this->getCell($coordinate)->setValue($value);
+        $cellAddress = Functions::trimSheetFromCellReference($this->validateCellAddress($coordinate));
+        $this->getCell($cellAddress)->setValue($value);
 
         return $this;
     }
 
     /**
      * Set a cell value by using numeric cell coordinates.
+     *
+     * @Deprecated 1.23.0
+     *      Use the setCellValue() method with a cell address such as 'C5' instead;,
+     *          or passing in an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
      *
      * @param int $columnIndex Numeric column coordinate of the cell
      * @param int $row Numeric row coordinate of the cell
@@ -1129,7 +1229,7 @@ class Worksheet implements IComparable
      */
     public function setCellValueByColumnAndRow($columnIndex, $row, $value)
     {
-        $this->getCellByColumnAndRow($columnIndex, $row)->setValue($value);
+        $this->getCell(Coordinate::stringFromColumnIndex($columnIndex) . $row)->setValue($value);
 
         return $this;
     }
@@ -1137,7 +1237,8 @@ class Worksheet implements IComparable
     /**
      * Set a cell value.
      *
-     * @param string $coordinate Coordinate of the cell, eg: 'A1'
+     * @param array<int>|CellAddress|string $coordinate Coordinate of the cell as a string, eg: 'C5';
+     *               or as an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
      * @param mixed $value Value of the cell
      * @param string $dataType Explicit data type, see DataType::TYPE_*
      *
@@ -1145,14 +1246,18 @@ class Worksheet implements IComparable
      */
     public function setCellValueExplicit($coordinate, $value, $dataType)
     {
-        // Set value
-        $this->getCell($coordinate)->setValueExplicit($value, $dataType);
+        $cellAddress = Functions::trimSheetFromCellReference($this->validateCellAddress($coordinate));
+        $this->getCell($cellAddress)->setValueExplicit($value, $dataType);
 
         return $this;
     }
 
     /**
      * Set a cell value by using numeric cell coordinates.
+     *
+     * @Deprecated 1.23.0
+     *      Use the setCellValueExplicit() method with a cell address such as 'C5' instead;,
+     *          or passing in an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
      *
      * @param int $columnIndex Numeric column coordinate of the cell
      * @param int $row Numeric row coordinate of the cell
@@ -1163,7 +1268,7 @@ class Worksheet implements IComparable
      */
     public function setCellValueExplicitByColumnAndRow($columnIndex, $row, $value, $dataType)
     {
-        $this->getCellByColumnAndRow($columnIndex, $row)->setValueExplicit($value, $dataType);
+        $this->getCell(Coordinate::stringFromColumnIndex($columnIndex) . $row)->setValueExplicit($value, $dataType);
 
         return $this;
     }
@@ -1171,22 +1276,25 @@ class Worksheet implements IComparable
     /**
      * Get cell at a specific coordinate.
      *
-     * @param string $coordinate Coordinate of the cell, eg: 'A1'
+     * @param array<int>|CellAddress|string $coordinate Coordinate of the cell as a string, eg: 'C5';
+     *               or as an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
      *
      * @return Cell Cell that was found or created
      */
-    public function getCell(string $coordinate): Cell
+    public function getCell($coordinate): Cell
     {
+        $cellAddress = Functions::trimSheetFromCellReference($this->validateCellAddress($coordinate));
+
         // Shortcut for increased performance for the vast majority of simple cases
-        if ($this->cellCollection->has($coordinate)) {
+        if ($this->cellCollection->has($cellAddress)) {
             /** @var Cell $cell */
-            $cell = $this->cellCollection->get($coordinate);
+            $cell = $this->cellCollection->get($cellAddress);
 
             return $cell;
         }
 
         /** @var Worksheet $sheet */
-        [$sheet, $finalCoordinate] = $this->getWorksheetAndCoordinate($coordinate);
+        [$sheet, $finalCoordinate] = $this->getWorksheetAndCoordinate($cellAddress);
         $cell = $sheet->cellCollection->get($finalCoordinate);
 
         return $cell ?? $sheet->createNewCell($finalCoordinate);
@@ -1210,7 +1318,7 @@ class Worksheet implements IComparable
             $sheet = $this->parent->getSheetByName($worksheetReference[0]);
             $finalCoordinate = strtoupper($worksheetReference[1]);
 
-            if (!$sheet) {
+            if ($sheet === null) {
                 throw new Exception('Sheet not found for name: ' . $worksheetReference[0]);
             }
         } elseif (
@@ -1221,7 +1329,7 @@ class Worksheet implements IComparable
             $namedRange = $this->validateNamedRange($coordinate, true);
             if ($namedRange !== null) {
                 $sheet = $namedRange->getWorksheet();
-                if (!$sheet) {
+                if ($sheet === null) {
                     throw new Exception('Sheet not found for named range: ' . $namedRange->getName());
                 }
 
@@ -1231,7 +1339,7 @@ class Worksheet implements IComparable
             }
         }
 
-        if (!$sheet || !$finalCoordinate) {
+        if ($sheet === null || $finalCoordinate === null) {
             $sheet = $this;
             $finalCoordinate = strtoupper($coordinate);
         }
@@ -1265,6 +1373,10 @@ class Worksheet implements IComparable
     /**
      * Get cell at a specific coordinate by using numeric cell coordinates.
      *
+     * @Deprecated 1.23.0
+     *      Use the getCell() method with a cell address such as 'C5' instead;,
+     *          or passing in an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
+     *
      * @param int $columnIndex Numeric column coordinate of the cell
      * @param int $row Numeric row coordinate of the cell
      *
@@ -1272,18 +1384,7 @@ class Worksheet implements IComparable
      */
     public function getCellByColumnAndRow($columnIndex, $row): Cell
     {
-        $columnLetter = Coordinate::stringFromColumnIndex($columnIndex);
-        $coordinate = $columnLetter . $row;
-
-        if ($this->cellCollection->has($coordinate)) {
-            /** @var Cell $cell */
-            $cell = $this->cellCollection->get($coordinate);
-
-            return $cell;
-        }
-
-        // Create new cell object, if required
-        return $this->createNewCell($coordinate);
+        return $this->getCell(Coordinate::stringFromColumnIndex($columnIndex) . $row);
     }
 
     /**
@@ -1328,14 +1429,14 @@ class Worksheet implements IComparable
     /**
      * Does the cell at a specific coordinate exist?
      *
-     * @param string $coordinate Coordinate of the cell eg: 'A1'
-     *
-     * @return bool
+     * @param array<int>|CellAddress|string $coordinate Coordinate of the cell as a string, eg: 'C5';
+     *               or as an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
      */
-    public function cellExists($coordinate)
+    public function cellExists($coordinate): bool
     {
+        $cellAddress = $this->validateCellAddress($coordinate);
         /** @var Worksheet $sheet */
-        [$sheet, $finalCoordinate] = $this->getWorksheetAndCoordinate($coordinate);
+        [$sheet, $finalCoordinate] = $this->getWorksheetAndCoordinate($cellAddress);
 
         return $sheet->cellCollection->has($finalCoordinate);
     }
@@ -1343,12 +1444,14 @@ class Worksheet implements IComparable
     /**
      * Cell at a specific coordinate by using numeric cell coordinates exists?
      *
+     * @Deprecated 1.23.0
+     *      Use the cellExists() method with a cell address such as 'C5' instead;,
+     *          or passing in an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
+     *
      * @param int $columnIndex Numeric column coordinate of the cell
      * @param int $row Numeric row coordinate of the cell
-     *
-     * @return bool
      */
-    public function cellExistsByColumnAndRow($columnIndex, $row)
+    public function cellExistsByColumnAndRow($columnIndex, $row): bool
     {
         return $this->cellExists(Coordinate::stringFromColumnIndex($columnIndex) . $row);
     }
@@ -1416,10 +1519,15 @@ class Worksheet implements IComparable
     /**
      * Get style for cell.
      *
-     * @param string $cellCoordinate Cell coordinate (or range) to get style for, eg: 'A1'
+     * @param AddressRange|array<int>|CellAddress|int|string $cellCoordinate
+     *              A simple string containing a cell address like 'A1' or a cell range like 'A1:E10'
+     *              or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *              or a CellAddress or AddressRange object.
      */
     public function getStyle($cellCoordinate): Style
     {
+        $cellCoordinate = $this->validateCellOrCellRange($cellCoordinate);
+
         // set this sheet as active
         $this->parent->setActiveSheetIndex($this->parent->getIndex($this));
 
@@ -1427,6 +1535,35 @@ class Worksheet implements IComparable
         $this->setSelectedCells($cellCoordinate);
 
         return $this->parent->getCellXfSupervisor();
+    }
+
+    /**
+     * Get style for cell by using numeric cell coordinates.
+     *
+     * @Deprecated 1.23.0
+     *      Use the getStyle() method with a cell address range such as 'C5:F8' instead;,
+     *          or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *          or an AddressRange object.
+     *
+     * @param int $columnIndex1 Numeric column coordinate of the cell
+     * @param int $row1 Numeric row coordinate of the cell
+     * @param null|int $columnIndex2 Numeric column coordinate of the range cell
+     * @param null|int $row2 Numeric row coordinate of the range cell
+     *
+     * @return Style
+     */
+    public function getStyleByColumnAndRow($columnIndex1, $row1, $columnIndex2 = null, $row2 = null)
+    {
+        if ($columnIndex2 !== null && $row2 !== null) {
+            $cellRange = new CellRange(
+                CellAddress::fromColumnAndRow($columnIndex1, $row1),
+                CellAddress::fromColumnAndRow($columnIndex2, $row2)
+            );
+
+            return $this->getStyle($cellRange);
+        }
+
+        return $this->getStyle(CellAddress::fromColumnAndRow($columnIndex1, $row1));
     }
 
     /**
@@ -1483,7 +1620,7 @@ class Worksheet implements IComparable
     {
         $coordinate = strtoupper($coordinate);
         if (strpos($coordinate, ':') !== false) {
-            return isset($this->conditionalStylesCollection[strtoupper($coordinate)]);
+            return isset($this->conditionalStylesCollection[$coordinate]);
         }
 
         $cell = $this->getCell($coordinate);
@@ -1533,27 +1670,6 @@ class Worksheet implements IComparable
         $this->conditionalStylesCollection[strtoupper($coordinate)] = $styles;
 
         return $this;
-    }
-
-    /**
-     * Get style for cell by using numeric cell coordinates.
-     *
-     * @param int $columnIndex1 Numeric column coordinate of the cell
-     * @param int $row1 Numeric row coordinate of the cell
-     * @param null|int $columnIndex2 Numeric column coordinate of the range cell
-     * @param null|int $row2 Numeric row coordinate of the range cell
-     *
-     * @return Style
-     */
-    public function getStyleByColumnAndRow($columnIndex1, $row1, $columnIndex2 = null, $row2 = null)
-    {
-        if ($columnIndex2 !== null && $row2 !== null) {
-            $cellRange = Coordinate::stringFromColumnIndex($columnIndex1) . $row1 . ':' . Coordinate::stringFromColumnIndex($columnIndex2) . $row2;
-
-            return $this->getStyle($cellRange);
-        }
-
-        return $this->getStyle(Coordinate::stringFromColumnIndex($columnIndex1) . $row1);
     }
 
     /**
@@ -1640,26 +1756,22 @@ class Worksheet implements IComparable
     /**
      * Set break on a cell.
      *
-     * @param string $coordinate Cell coordinate (e.g. A1)
+     * @param array<int>|CellAddress|string $coordinate Coordinate of the cell as a string, eg: 'C5';
+     *               or as an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
      * @param int $break Break type (type of Worksheet::BREAK_*)
      *
      * @return $this
      */
     public function setBreak($coordinate, $break)
     {
-        // Uppercase coordinate
-        $coordinate = strtoupper($coordinate);
+        $cellAddress = Functions::trimSheetFromCellReference($this->validateCellAddress($coordinate));
 
-        if ($coordinate != '') {
-            if ($break == self::BREAK_NONE) {
-                if (isset($this->breaks[$coordinate])) {
-                    unset($this->breaks[$coordinate]);
-                }
-            } else {
-                $this->breaks[$coordinate] = $break;
+        if ($break === self::BREAK_NONE) {
+            if (isset($this->breaks[$cellAddress])) {
+                unset($this->breaks[$cellAddress]);
             }
         } else {
-            throw new Exception('No cell coordinate specified.');
+            $this->breaks[$cellAddress] = $break;
         }
 
         return $this;
@@ -1667,6 +1779,10 @@ class Worksheet implements IComparable
 
     /**
      * Set break on a cell by using numeric cell coordinates.
+     *
+     * @Deprecated 1.23.0
+     *      Use the setBreak() method with a cell address such as 'C5' instead;,
+     *          or passing in an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
      *
      * @param int $columnIndex Numeric column coordinate of the cell
      * @param int $row Numeric row coordinate of the cell
@@ -1692,18 +1808,15 @@ class Worksheet implements IComparable
     /**
      * Set merge on a cell range.
      *
-     * @param string $range Cell range (e.g. A1:E1)
+     * @param AddressRange|array<int>|string $range A simple string containing a Cell range like 'A1:E10'
+     *              or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *              or an AddressRange.
      *
      * @return $this
      */
     public function mergeCells($range)
     {
-        // Uppercase coordinate
-        $range = strtoupper($range);
-        // Convert 'A:C' to 'A1:C1048576'
-        $range = self::pregReplace('/^([A-Z]+):([A-Z]+)$/', '${1}1:${2}1048576', $range);
-        // Convert '1:3' to 'A1:XFD3'
-        $range = self::pregReplace('/^(\\d+):(\\d+)$/', 'A${1}:XFD${2}', $range);
+        $range = Functions::trimSheetFromCellReference($this->validateCellRange($range));
 
         if (preg_match('/^([A-Z]+)(\\d+):([A-Z]+)(\\d+)$/', $range, $matches) === 1) {
             $this->mergeCells[$range] = $range;
@@ -1717,7 +1830,7 @@ class Worksheet implements IComparable
             $numberColumns = $lastColumnIndex - $firstColumnIndex;
 
             // create upper left cell if it does not already exist
-            $upperLeft = "$firstColumn$firstRow";
+            $upperLeft = "{$firstColumn}{$firstRow}";
             if (!$this->cellExists($upperLeft)) {
                 $this->getCell($upperLeft)->setValueExplicit(null, DataType::TYPE_NULL);
             }
@@ -1779,6 +1892,11 @@ class Worksheet implements IComparable
     /**
      * Set merge on a cell range by using numeric cell coordinates.
      *
+     * @Deprecated 1.23.0
+     *      Use the mergeCells() method with a cell address range such as 'C5:F8' instead;,
+     *          or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *          or an AddressRange object.
+     *
      * @param int $columnIndex1 Numeric column coordinate of the first cell
      * @param int $row1 Numeric row coordinate of the first cell
      * @param int $columnIndex2 Numeric column coordinate of the last cell
@@ -1788,7 +1906,10 @@ class Worksheet implements IComparable
      */
     public function mergeCellsByColumnAndRow($columnIndex1, $row1, $columnIndex2, $row2)
     {
-        $cellRange = Coordinate::stringFromColumnIndex($columnIndex1) . $row1 . ':' . Coordinate::stringFromColumnIndex($columnIndex2) . $row2;
+        $cellRange = new CellRange(
+            CellAddress::fromColumnAndRow($columnIndex1, $row1),
+            CellAddress::fromColumnAndRow($columnIndex2, $row2)
+        );
 
         return $this->mergeCells($cellRange);
     }
@@ -1796,14 +1917,15 @@ class Worksheet implements IComparable
     /**
      * Remove merge on a cell range.
      *
-     * @param string $range Cell range (e.g. A1:E1)
+     * @param AddressRange|array<int>|string $range A simple string containing a Cell range like 'A1:E10'
+     *              or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *              or an AddressRange.
      *
      * @return $this
      */
     public function unmergeCells($range)
     {
-        // Uppercase coordinate
-        $range = strtoupper($range);
+        $range = Functions::trimSheetFromCellReference($this->validateCellRange($range));
 
         if (strpos($range, ':') !== false) {
             if (isset($this->mergeCells[$range])) {
@@ -1821,6 +1943,11 @@ class Worksheet implements IComparable
     /**
      * Remove merge on a cell range by using numeric cell coordinates.
      *
+     * @Deprecated 1.23.0
+     *      Use the unmergeCells() method with a cell address range such as 'C5:F8' instead;,
+     *          or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *          or an AddressRange object.
+     *
      * @param int $columnIndex1 Numeric column coordinate of the first cell
      * @param int $row1 Numeric row coordinate of the first cell
      * @param int $columnIndex2 Numeric column coordinate of the last cell
@@ -1830,7 +1957,10 @@ class Worksheet implements IComparable
      */
     public function unmergeCellsByColumnAndRow($columnIndex1, $row1, $columnIndex2, $row2)
     {
-        $cellRange = Coordinate::stringFromColumnIndex($columnIndex1) . $row1 . ':' . Coordinate::stringFromColumnIndex($columnIndex2) . $row2;
+        $cellRange = new CellRange(
+            CellAddress::fromColumnAndRow($columnIndex1, $row1),
+            CellAddress::fromColumnAndRow($columnIndex2, $row2)
+        );
 
         return $this->unmergeCells($cellRange);
     }
@@ -1861,9 +1991,11 @@ class Worksheet implements IComparable
     }
 
     /**
-     * Set protection on a cell range.
+     * Set protection on a cell or cell range.
      *
-     * @param string $range Cell (e.g. A1) or cell range (e.g. A1:E1)
+     * @param AddressRange|array<int>|CellAddress|int|string $range A simple string containing a Cell range like 'A1:E10'
+     *              or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *              or a CellAddress or AddressRange object.
      * @param string $password Password to unlock the protection
      * @param bool $alreadyHashed If the password has already been hashed, set this to true
      *
@@ -1871,8 +2003,7 @@ class Worksheet implements IComparable
      */
     public function protectCells($range, $password, $alreadyHashed = false)
     {
-        // Uppercase coordinate
-        $range = strtoupper($range);
+        $range = Functions::trimSheetFromCellReference($this->validateCellOrCellRange($range));
 
         if (!$alreadyHashed) {
             $password = Shared\PasswordHasher::hashPassword($password);
@@ -1885,6 +2016,11 @@ class Worksheet implements IComparable
     /**
      * Set protection on a cell range by using numeric cell coordinates.
      *
+     * @Deprecated 1.23.0
+     *      Use the protectCells() method with a cell address range such as 'C5:F8' instead;,
+     *          or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *          or an AddressRange object.
+     *
      * @param int $columnIndex1 Numeric column coordinate of the first cell
      * @param int $row1 Numeric row coordinate of the first cell
      * @param int $columnIndex2 Numeric column coordinate of the last cell
@@ -1896,22 +2032,26 @@ class Worksheet implements IComparable
      */
     public function protectCellsByColumnAndRow($columnIndex1, $row1, $columnIndex2, $row2, $password, $alreadyHashed = false)
     {
-        $cellRange = Coordinate::stringFromColumnIndex($columnIndex1) . $row1 . ':' . Coordinate::stringFromColumnIndex($columnIndex2) . $row2;
+        $cellRange = new CellRange(
+            CellAddress::fromColumnAndRow($columnIndex1, $row1),
+            CellAddress::fromColumnAndRow($columnIndex2, $row2)
+        );
 
         return $this->protectCells($cellRange, $password, $alreadyHashed);
     }
 
     /**
-     * Remove protection on a cell range.
+     * Remove protection on a cell or cell range.
      *
-     * @param string $range Cell (e.g. A1) or cell range (e.g. A1:E1)
+     * @param AddressRange|array<int>|CellAddress|int|string $range A simple string containing a Cell range like 'A1:E10'
+     *              or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *              or a CellAddress or AddressRange object.
      *
      * @return $this
      */
     public function unprotectCells($range)
     {
-        // Uppercase coordinate
-        $range = strtoupper($range);
+        $range = Functions::trimSheetFromCellReference($this->validateCellOrCellRange($range));
 
         if (isset($this->protectedCells[$range])) {
             unset($this->protectedCells[$range]);
@@ -1925,6 +2065,11 @@ class Worksheet implements IComparable
     /**
      * Remove protection on a cell range by using numeric cell coordinates.
      *
+     * @Deprecated 1.23.0
+     *      Use the protectCells() method with a cell address range such as 'C5:F8' instead;,
+     *          or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *          or an AddressRange object.
+     *
      * @param int $columnIndex1 Numeric column coordinate of the first cell
      * @param int $row1 Numeric row coordinate of the first cell
      * @param int $columnIndex2 Numeric column coordinate of the last cell
@@ -1934,7 +2079,10 @@ class Worksheet implements IComparable
      */
     public function unprotectCellsByColumnAndRow($columnIndex1, $row1, $columnIndex2, $row2)
     {
-        $cellRange = Coordinate::stringFromColumnIndex($columnIndex1) . $row1 . ':' . Coordinate::stringFromColumnIndex($columnIndex2) . $row2;
+        $cellRange = new CellRange(
+            CellAddress::fromColumnAndRow($columnIndex1, $row1),
+            CellAddress::fromColumnAndRow($columnIndex2, $row2)
+        );
 
         return $this->unprotectCells($cellRange);
     }
@@ -1962,17 +2110,21 @@ class Worksheet implements IComparable
     /**
      * Set AutoFilter.
      *
-     * @param AutoFilter|string $autoFilterOrRange
+     * @param AddressRange|array<int>|AutoFilter|string $autoFilterOrRange
      *            A simple string containing a Cell range like 'A1:E10' is permitted for backward compatibility
+     *              or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *              or an AddressRange.
      *
      * @return $this
      */
     public function setAutoFilter($autoFilterOrRange)
     {
-        if (is_string($autoFilterOrRange)) {
-            $this->autoFilter->setRange($autoFilterOrRange);
-        } elseif (is_object($autoFilterOrRange) && ($autoFilterOrRange instanceof AutoFilter)) {
+        if (is_object($autoFilterOrRange) && ($autoFilterOrRange instanceof AutoFilter)) {
             $this->autoFilter = $autoFilterOrRange;
+        } else {
+            $cellRange = Functions::trimSheetFromCellReference($this->validateCellRange($autoFilterOrRange));
+
+            $this->autoFilter->setRange($cellRange);
         }
 
         return $this;
@@ -1980,6 +2132,11 @@ class Worksheet implements IComparable
 
     /**
      * Set Autofilter Range by using numeric cell coordinates.
+     *
+     * @Deprecated 1.23.0
+     *      Use the setAutoFilter() method with a cell address range such as 'C5:F8' instead;,
+     *          or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *          or an AddressRange object or AutoFilter object.
      *
      * @param int $columnIndex1 Numeric column coordinate of the first cell
      * @param int $row1 Numeric row coordinate of the first cell
@@ -1990,11 +2147,12 @@ class Worksheet implements IComparable
      */
     public function setAutoFilterByColumnAndRow($columnIndex1, $row1, $columnIndex2, $row2)
     {
-        return $this->setAutoFilter(
-            Coordinate::stringFromColumnIndex($columnIndex1) . $row1
-            . ':' .
-            Coordinate::stringFromColumnIndex($columnIndex2) . $row2
+        $cellRange = new CellRange(
+            CellAddress::fromColumnAndRow($columnIndex1, $row1),
+            CellAddress::fromColumnAndRow($columnIndex2, $row2)
         );
+
+        return $this->setAutoFilter($cellRange);
     }
 
     /**
@@ -2026,23 +2184,33 @@ class Worksheet implements IComparable
      *     - B1 will freeze the columns to the left of cell B1 (i.e column A)
      *     - B2 will freeze the rows above and to the left of cell B2 (i.e row 1 and column A)
      *
-     * @param null|string $cell Position of the split
-     * @param null|string $topLeftCell default position of the right bottom pane
+     * @param null|array<int>|CellAddress|string $coordinate Coordinate of the cell as a string, eg: 'C5';
+     *            or as an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
+     *        Passing a null value for this argument will clear any existing freeze pane for this worksheet.
+     * @param null|array<int>|CellAddress|string $topLeftCell default position of the right bottom pane
+     *            Coordinate of the cell as a string, eg: 'C5'; or as an array of [$columnIndex, $row] (e.g. [3, 5]),
+     *            or a CellAddress object.
      *
      * @return $this
      */
-    public function freezePane($cell, $topLeftCell = null)
+    public function freezePane($coordinate, $topLeftCell = null)
     {
-        if (is_string($cell) && Coordinate::coordinateIsRange($cell)) {
+        $cellAddress = ($coordinate !== null)
+            ? Functions::trimSheetFromCellReference($this->validateCellAddress($coordinate))
+            : null;
+        if ($cellAddress !== null && Coordinate::coordinateIsRange($cellAddress)) {
             throw new Exception('Freeze pane can not be set on a range of cells.');
         }
+        $topLeftCell = ($topLeftCell !== null)
+            ? Functions::trimSheetFromCellReference($this->validateCellAddress($topLeftCell))
+            : null;
 
-        if ($cell !== null && $topLeftCell === null) {
-            $coordinate = Coordinate::coordinateFromString($cell);
+        if ($cellAddress !== null && $topLeftCell === null) {
+            $coordinate = Coordinate::coordinateFromString($cellAddress);
             $topLeftCell = $coordinate[0] . $coordinate[1];
         }
 
-        $this->freezePane = $cell;
+        $this->freezePane = $cellAddress;
         $this->topLeftCell = $topLeftCell;
 
         return $this;
@@ -2057,6 +2225,10 @@ class Worksheet implements IComparable
 
     /**
      * Freeze Pane by using numeric cell coordinates.
+     *
+     * @Deprecated 1.23.0
+     *      Use the freezePane() method with a cell address such as 'C5' instead;,
+     *          or passing in an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
      *
      * @param int $columnIndex Numeric column coordinate of the cell
      * @param int $row Numeric row coordinate of the cell
@@ -2427,37 +2599,41 @@ class Worksheet implements IComparable
     /**
      * Get comment for cell.
      *
-     * @param string $cellCoordinate Cell coordinate to get comment for, eg: 'A1'
+     * @param array<int>|CellAddress|string $cellCoordinate Coordinate of the cell as a string, eg: 'C5';
+     *               or as an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
      *
      * @return Comment
      */
     public function getComment($cellCoordinate)
     {
-        // Uppercase coordinate
-        $cellCoordinate = strtoupper($cellCoordinate);
+        $cellAddress = Functions::trimSheetFromCellReference($this->validateCellAddress($cellCoordinate));
 
-        if (Coordinate::coordinateIsRange($cellCoordinate)) {
+        if (Coordinate::coordinateIsRange($cellAddress)) {
             throw new Exception('Cell coordinate string can not be a range of cells.');
-        } elseif (strpos($cellCoordinate, '$') !== false) {
+        } elseif (strpos($cellAddress, '$') !== false) {
             throw new Exception('Cell coordinate string must not be absolute.');
-        } elseif ($cellCoordinate == '') {
+        } elseif ($cellAddress == '') {
             throw new Exception('Cell coordinate can not be zero-length string.');
         }
 
         // Check if we already have a comment for this cell.
-        if (isset($this->comments[$cellCoordinate])) {
-            return $this->comments[$cellCoordinate];
+        if (isset($this->comments[$cellAddress])) {
+            return $this->comments[$cellAddress];
         }
 
         // If not, create a new comment.
         $newComment = new Comment();
-        $this->comments[$cellCoordinate] = $newComment;
+        $this->comments[$cellAddress] = $newComment;
 
         return $newComment;
     }
 
     /**
      * Get comment for cell by using numeric cell coordinates.
+     *
+     * @Deprecated 1.23.0
+     *      Use the getComment() method with a cell address such as 'C5' instead;,
+     *          or passing in an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
      *
      * @param int $columnIndex Numeric column coordinate of the cell
      * @param int $row Numeric row coordinate of the cell
@@ -2517,48 +2693,21 @@ class Worksheet implements IComparable
         return self::ensureString(preg_replace($pattern, $replacement, $subject));
     }
 
-    private function tryDefinedName(string $coordinate): string
-    {
-        // Uppercase coordinate
-        $coordinate = strtoupper($coordinate);
-        // Eliminate leading equal sign
-        $coordinate = self::pregReplace('/^=/', '', $coordinate);
-        $defined = $this->parent->getDefinedName($coordinate, $this);
-        if ($defined !== null) {
-            if ($defined->getWorksheet() === $this && !$defined->isFormula()) {
-                $coordinate = self::pregReplace('/^=/', '', $defined->getValue());
-            }
-        }
-
-        return $coordinate;
-    }
-
     /**
      * Select a range of cells.
      *
-     * @param string $coordinate Cell range, examples: 'A1', 'B2:G5', 'A:C', '3:6'
+     * @param AddressRange|array<int>|CellAddress|int|string $coordinate A simple string containing a Cell range like 'A1:E10'
+     *              or passing in an array of [$fromColumnIndex, $fromRow, $toColumnIndex, $toRow] (e.g. [3, 5, 6, 8]),
+     *              or a CellAddress or AddressRange object.
      *
      * @return $this
      */
     public function setSelectedCells($coordinate)
     {
-        $originalCoordinate = $coordinate;
-        $coordinate = $this->tryDefinedName($coordinate);
-
-        // Convert 'A' to 'A:A'
-        $coordinate = self::pregReplace('/^([A-Z]+)$/', '${1}:${1}', $coordinate);
-
-        // Convert '1' to '1:1'
-        $coordinate = self::pregReplace('/^(\d+)$/', '${1}:${1}', $coordinate);
-
-        // Convert 'A:C' to 'A1:C1048576'
-        $coordinate = self::pregReplace('/^([A-Z]+):([A-Z]+)$/', '${1}1:${2}1048576', $coordinate);
-
-        // Convert '1:3' to 'A1:XFD3'
-        $coordinate = self::pregReplace('/^(\d+):(\d+)$/', 'A${1}:XFD${2}', $coordinate);
-        if (preg_match('/^\\$?[A-Z]{1,3}\\$?\d{1,7}(:\\$?[A-Z]{1,3}\\$?\d{1,7})?$/', $coordinate) !== 1) {
-            throw new Exception("Invalid setSelectedCells $originalCoordinate $coordinate");
+        if (is_string($coordinate)) {
+            $coordinate = $this->tryDefinedName($coordinate);
         }
+        $coordinate = $this->validateCellOrCellRange($coordinate);
 
         if (Coordinate::coordinateIsRange($coordinate)) {
             [$first] = Coordinate::splitRange($coordinate);
@@ -2573,6 +2722,10 @@ class Worksheet implements IComparable
 
     /**
      * Selected cell by using numeric cell coordinates.
+     *
+     * @Deprecated 1.23.0
+     *      Use the setSelectedCells() method with a cell address such as 'C5' instead;,
+     *          or passing in an array of [$columnIndex, $row] (e.g. [3, 5]), or a CellAddress object.
      *
      * @param int $columnIndex Numeric column coordinate of the cell
      * @param int $row Numeric row coordinate of the cell